+++ src/main/java/racingcar/domain/car/Car.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.domain.car;
 
 public class Car implements Comparable<Car> {
     private Name name;
