+++ src/main/java/racingcar/view/ResultView.java
@@ -1,7 +1,7 @@
 package racingcar.view;
 
-import racingcar.car.Car;
-import racingcar.car.Cars;
+import racingcar.domain.car.Car;
+import racingcar.domain.car.Cars;
 
 import java.util.List;
 import java.util.StringJoiner;
